requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-extensions==3.2.3
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
psycopg2-binary==2.9.7
pytz==2025.2
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
